requirements.txt
Django==5.2
djangorestframework==3.16.1
psycopg2==2.9.11
drf-spectacular==0.29.0
Pillow==12.0.0
uWSGI==2.0.31
gunicorn==23.0.0